requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpatch==1.33
langchain-core==0.3.34
langchain-text-splitters==0.3.6
langsmith==0.3.7
pydantic==2.10.6
requests==2.32.3
requests-toolbelt==1.0.0
SQLAlchemy==2.0.38
yarl==1.18.3
langchain==0.3.16
python-dotenv==1.0.1
langchain-groq==0.2.4
pandas
langchain-community==0.3.16
pypdf==4.3.1
wikipedia==1.4.0
bs4==0.0.2
unstructured
langchain-ollama
langchain-huggingface
langchain-chroma
faiss-cpu
pandas
langchain-groq
google-generativeai
langchain-google-genai
langsmith
langgraph
langserve
FastAPI
Uvicorn
sse_starlette
streamlit
hdbcli
